gpt-researcher/backend/requirements.txt
# Backend-specific requirements
# For production backend deployment

# Core Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.1
python-dotenv>=1.0.0
websockets>=13.1
python-multipart>=0.0.6
Pillow>=10.0.0

# LangChain v1
langchain>=1.0.0
langchain-classic>=1.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-text-splitters>=1.0.0

# LLM & API
openai>=1.3.3
httpx>=0.28.1
tavily-python>=0.7.12

# Output formats
aiofiles>=23.2.1
mistune>=3.0.2
md2pdf>=1.0.1
python-docx>=1.1.0
htmldocx>=0.0.6
jinja2>=3.1.6

# GPT-Researcher (install from root)
# Run: pip install -e . from the project root
# gpt-researcher>=0.14.4
